+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KMYZZK7QTD2683VANAFXM615"
+        "01KMZ48P7EES7MWB6ZWJAWHV39"
     )
     r"""The unique identifier of the data source"""