+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasetop.py
@@ -154,7 +154,7 @@ class UpdateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-03-30T09:07:30.737Z")
+    updated: Optional[datetime] = parse_datetime("2026-03-30T10:22:25.091Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")