+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KMYZZK7V59HSG0SG245XR69E"
+        "01KMZ48P7RVRRC4TRZ1MQQC35Y"
     )
     r"""The unique identifier of the data source"""